:fire: Remove redundant code
